started creating utils - independent helper functions
